moved atlas up
