+++ gameserver/src/test/java/com/lhf/game/map/KeyedDoorwayTest.java
@@ -7,10 +7,10 @@ import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
 
 import com.google.common.truth.Truth;
+import com.lhf.game.Atlas.AtlasMappingItem;
+import com.lhf.game.Atlas.TargetedTester;
 import com.lhf.game.creature.intelligence.AIComBundle;
 import com.lhf.game.item.concrete.LockKey;
-import com.lhf.game.map.Atlas.AtlasMappingItem;
-import com.lhf.game.map.Atlas.TargetedTester;
 import com.lhf.game.map.Dungeon.DungeonBuilder;
 import com.lhf.game.map.Room.RoomBuilder;
 import com.lhf.messages.GameEventType;
